Revert "liblink, cmd/5a: fix reglist parsing/printing after changing REG_R0 to 32"
This reverts commit e80ad808c4f15a5c2a68009cf00b0d3b38274e02.

Reviewed-on: https://go-review.googlesource.com/3960

+++ src/cmd/5a/y.tab.h
@@ -149,7 +149,7 @@ typedef union YYSTYPE
 #line 39 "a.y"
 {
 	Sym	*sym;
-	int64	lval;
+	int32	lval;
 	double	dval;
 	char	sval[8];
 	Addr	addr;
